Improve Loading System Performance (#112)

+++ Source/Core/Loader/ERS_ModelLoader/ERS_ModelLoader.cpp
@@ -184,8 +184,8 @@ void ERS_CLASS_ModelLoader::ProcessGPU(std::shared_ptr<ERS_STRUCT_Model> Model) 
     }
 
     // Process Texture References, Setup Meshes
-    for (int i = 0; (long)i < (long)Model->Meshes.size(); i++) {
-        for (int Index = 0; (long)Index < (long)Model->Meshes[i].TextureReferences_.size(); Index++) { // IF TEXTURES DONT WORK, CHECK HERE!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+    for (unsigned long i = 0; i < Model->Meshes.size(); i++) {
+        for (unsigned long Index = 0; Index < Model->Meshes[i].TextureReferences_.size(); Index++) { // IF TEXTURES DONT WORK, CHECK HERE!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
             int TextureIndex = Model->Meshes[i].TextureReferences_[Index];
